tacky: refactoring into smaller modules
