+++ src/tacky/expr.rs
@@ -1,774 +1,9 @@
-use crate::ast;
-use crate::sema;
-use std::collections::HashMap;
-use std::rc::Rc;
-
-#[derive(Debug)]
-pub struct Program {
-    pub top_level: Vec<TopLevel>,
-    pub symbols: SymbolTable,
-}
-
-#[derive(Debug, PartialEq)]
-pub enum TopLevel {
-    Fun(Function),
-    Static(StaticVariable),
-}
-
-#[derive(Debug, Default)]
-pub struct SymbolTable {
-    pub table: HashMap<Rc<String>, SymbolEntry>,
-}
-
-impl SymbolTable {
-    pub fn get(&self, key: &Rc<String>) -> Option<&SymbolEntry> {
-        self.table.get(key)
-    }
-
-    pub fn new_entry(&mut self, key: Rc<String>, r#type: ast::Type) {
-        let old_key = self.table.insert(
-            Rc::clone(&key),
-            SymbolEntry {
-                r#type,
-                attribute: sema::tc::Attribute::Local,
-            },
-        );
-
-        assert!(
-            old_key.is_none(),
-            "Every new entry into SymbolTable should have a unique name!, but {key} did not!"
-        );
-    }
-}
-
-#[derive(Clone, Debug)]
-pub struct SymbolEntry {
-    pub r#type: ast::Type,
-    pub attribute: sema::tc::Attribute,
-}
-
-impl From<sema::tc::SymbolTable> for SymbolTable {
-    fn from(value: sema::tc::SymbolTable) -> Self {
-        Self {
-            table: value
-                .global
-                .into_iter()
-                .map(|(k, v)| {
-                    (
-                        k,
-                        SymbolEntry {
-                            r#type: v.r#type,
-                            attribute: v.attribute,
-                        },
-                    )
-                })
-                .collect(),
-        }
-    }
-}
-
-impl StaticVariable {
-    fn from_symbol_with_name(name: Rc<String>, symbol: &sema::tc::SymbolEntry) -> Option<Self> {
-        match &symbol.attribute {
-            sema::tc::Attribute::Fun { .. } => None,
-            sema::tc::Attribute::Static {
-                initial_value,
-                external_linkage,
-            } => match initial_value {
-                sema::tc::InitialValue::Initial(i) => Some(StaticVariable {
-                    identifier: name,
-                    external_linkage: *external_linkage,
-                    init: i.to_vec(),
-                }),
-                sema::tc::InitialValue::Tentative => Some(StaticVariable {
-                    identifier: name,
-                    external_linkage: *external_linkage,
-                    init: vec![vec![0; symbol.r#type.base.nbytes()].into()],
-                }),
-                sema::tc::InitialValue::None => None,
-            },
-            sema::tc::Attribute::Local => None,
-        }
-    }
-}
-
-#[derive(Debug, PartialEq)]
-pub struct StaticVariable {
-    pub identifier: Rc<String>,
-    pub external_linkage: bool,
-    pub init: Vec<Rc<[u8]>>,
-}
-
-impl From<sema::ValidAst> for Program {
-    fn from(ast: sema::ValidAst) -> Self {
-        let sema::ValidAst { program, symbols } = ast;
-        let mut statics = vec![];
-        for (name, symbol) in symbols.global.iter() {
-            if let Some(r#static) = StaticVariable::from_symbol_with_name(Rc::clone(name), symbol) {
-                statics.push(r#static);
-            }
-        }
-        let mut symbols = SymbolTable::from(symbols);
-        let mut valid_functions = vec![];
-        for decl in program.declarations.into_iter() {
-            match decl {
-                // Only declarations with bodies will be returned here.
-                // We need to do some fixup so that if the definition for a
-                // function was not marked static but the first declaration was
-                // that the function gets defined as static.
-                ast::Declaration::FunDecl(f) => {
-                    if let Some(f) = Function::from_symbol(f, &mut symbols) {
-                        valid_functions.push(f);
-                    }
-                }
-                ast::Declaration::VarDecl(_) => {}
-            };
-        }
-        let top_level = valid_functions
-            .into_iter()
-            .map(TopLevel::Fun)
-            .chain(statics.into_iter().map(TopLevel::Static))
-            .collect::<Vec<TopLevel>>();
-        Self { top_level, symbols }
-    }
-}
-
-#[derive(Debug, PartialEq)]
-pub struct Function {
-    pub name: Rc<String>,
-    pub external_linkage: bool,
-    pub signature: Vec<(ast::Type, Option<Rc<String>>)>,
-    pub instructions: Vec<Instruction>,
-}
-
-impl Function {
-    fn make_temp_var(name: Rc<String>, counter: &'_ mut usize) -> impl FnMut() -> String + use<'_> {
-        move || {
-            let n = *counter;
-            *counter += 1;
-            format!("tacky.{name}.{n}")
-        }
-    }
-
-    fn make_tacky_temp_var(
-        r#type: ast::Type,
-        symbols: &mut SymbolTable,
-        make_temp_var: &mut impl FnMut() -> String,
-    ) -> Val {
-        let name = Rc::new(make_temp_var());
-        symbols.new_entry(Rc::clone(&name), r#type);
-        Val::Var(name)
-    }
-}
-
-impl Function {
-    fn from_symbol(decl: ast::FunDecl, symbols: &mut SymbolTable) -> Option<Self> {
-        // Insert function parameter types for type inference
-        // FIXME: Make sure that parameter names are unique!!!!
-        let mut signature = vec![];
-        for (r#type, name) in decl
-            .signature()
-            .expect("tacky.Function.from_symbol(): Error getting function declaration signature.")
-            .into_iter()
-        {
-            if let Some(name) = name {
-                symbols.new_entry(Rc::clone(name), r#type.clone());
-                signature.push((r#type.clone(), Some(Rc::clone(name))));
-            } else {
-                signature.push((r#type.clone(), None));
-            }
-        }
-
-        let ast::FunDecl {
-            name,
-            block: Some(block),
-            ..
-        } = decl
-        else {
-            return None;
-        };
-
-        let mut temp_var_counter = 0;
-        let mut make_temp_var =
-            Function::make_temp_var(Rc::new(name.to_string()), &mut temp_var_counter);
-        let mut instructions = Instruction::parse_block_with(block, symbols, &mut make_temp_var);
-
-        // Temporary fix suggested by the book for the case where a function
-        // is supposed to return something but does not.
-        instructions.push(Instruction::Return(Some(Val::Constant(
-            ast::Constant::I32(0),
-        ))));
-
-        let name = Rc::new(name.to_string());
-
-        // Check symbol table to get external linkage since the function
-        // declaration could be static but the definition can elide it.
-        // ```
-        // static int foo();
-        // int foo() {
-        //      ...
-        // }
-        // ```
-
-        let external_linkage = {
-            if let Some(SymbolEntry {
-                r#type:
-                    ast::Type {
-                        base: ast::BaseType::Fun { .. },
-                        ..
-                    },
-                attribute,
-            }) = symbols.get(&name)
-            {
-                attribute.has_external_linkage()
-            } else {
-                unreachable!()
-            }
-        };
-
-        Some(Function {
-            name,
-            signature,
-            external_linkage,
-            instructions,
-        })
-    }
-}
-
-#[derive(Debug, PartialEq)]
-pub enum Instruction {
-    Return(Option<Val>),
-    SignExtend {
-        src: Val,
-        dst: Val,
-    },
-    ZeroExtend {
-        src: Val,
-        dst: Val,
-    },
-    DoubleToInt {
-        src: Val,
-        dst: Val,
-    },
-    IntToDouble {
-        src: Val,
-        dst: Val,
-    },
-    DoubleToUInt {
-        src: Val,
-        dst: Val,
-    },
-    UIntToDouble {
-        src: Val,
-        dst: Val,
-    },
-    Truncate {
-        src: Val,
-        dst: Val,
-    },
-    Unary {
-        op: UnaryOp,
-        src: Val,
-        dst: Val,
-    },
-    Binary {
-        op: BinaryOp,
-        src1: Val,
-        src2: Val,
-        dst: Val,
-    },
-    Copy {
-        src: Val,
-        dst: Val,
-    },
-    GetAddress {
-        src: Val,
-        dst: Val,
-    },
-    Load {
-        src_ptr: Val,
-        dst: Val,
-    },
-    Store {
-        src: Val,
-        dst_ptr: Val,
-    },
-    Jump(Rc<String>),
-    JumpIfZero {
-        condition: Val,
-        target: Rc<String>,
-    },
-    JumpIfNotZero {
-        condition: Val,
-        target: Rc<String>,
-    },
-    Label(Rc<String>),
-    FunCall {
-        name: Rc<String>,
-        args: Vec<Val>,
-        dst: Val,
-    },
-    AddPtr {
-        ptr: Val,
-        index: Val,
-        scale: usize,
-        dst: Val,
-    },
-    CopyToOffset {
-        src: Val,
-        dst: Rc<String>,
-        offset: isize,
-    },
-}
-
-impl Instruction {
-    fn parse_decl_with(
-        decl: ast::Declaration,
-        symbols: &mut SymbolTable,
-        make_temp_var: &mut impl FnMut() -> String,
-    ) -> Vec<Self> {
-        match decl {
-            ast::Declaration::VarDecl(decl) => {
-                Self::parse_var_decl_with(decl, symbols, make_temp_var)
-            }
-            ast::Declaration::FunDecl(decl) => Self::parse_fun_decl_with(decl, make_temp_var),
-        }
-    }
-
-    fn parse_fun_decl_with(
-        decl: ast::FunDecl,
-        _make_temp_var: &mut impl FnMut() -> String,
-    ) -> Vec<Self> {
-        assert!(decl.block.is_none());
-        vec![]
-    }
-
-    fn process_initializer_rec(
-        base: usize,
-        in_array: bool,
-        name: Rc<String>,
-        init: ast::Initializer,
-        r#type: &ast::Type,
-        symbols: &mut SymbolTable,
-        make_temp_var: &mut impl FnMut() -> String,
-    ) -> Vec<Self> {
-        match init {
-            ast::Initializer::SingleInit(init) => {
-                let Expr {
-                    mut instructions,
-                    val: src,
-                } = Expr::parse_with_and_convert(*init, symbols, make_temp_var);
-                if in_array {
-                    instructions.push(Instruction::CopyToOffset {
-                        src,
-                        dst: Rc::clone(&name),
-                        offset: base.try_into().unwrap(),
-                    });
-                } else {
-                    let dst = Val::Var(name);
-                    instructions.push(Instruction::Copy {
-                        src,
-                        dst: dst.clone(),
-                    });
-                }
-                instructions
-            }
-            ast::Initializer::CompundInit(inits) => {
-                let mut instructions = vec![];
-                let per_element_size = r#type.base.size_of_base_type();
-                let mut current_base = base;
-                for init in inits {
-                    instructions.extend(Self::process_initializer_rec(
-                        current_base,
-                        true,
-                        name.clone(),
-                        init,
-                        r#type,
-                        symbols,
-                        make_temp_var,
-                    ));
-                    current_base += per_element_size;
-                }
-                instructions
-            }
-        }
-    }
-
-    fn process_initializer(
-        name: Rc<String>,
-        init: ast::Initializer,
-        r#type: &ast::Type,
-        symbols: &mut SymbolTable,
-        make_temp_var: &mut impl FnMut() -> String,
-    ) -> Vec<Self> {
-        Self::process_initializer_rec(0, false, name, init, r#type, symbols, make_temp_var)
-    }
-
-    fn parse_var_decl_with(
-        decl: ast::VarDecl,
-        symbols: &mut SymbolTable,
-        make_temp_var: &mut impl FnMut() -> String,
-    ) -> Vec<Self> {
-        if decl.storage_class != Some(ast::StorageClass::Extern) {
-            symbols.new_entry(Rc::clone(&decl.name), decl.r#type.clone());
-        }
-        match decl.init {
-            Some(init) => Self::process_initializer(
-                Rc::clone(&decl.name),
-                init,
-                &decl.r#type,
-                symbols,
-                make_temp_var,
-            ),
-            _ => vec![],
-        }
-    }
-
-    fn parse_stmt_with(
-        stmt: ast::Stmt,
-        symbols: &mut SymbolTable,
-        make_temp_var: &mut impl FnMut() -> String,
-    ) -> Vec<Self> {
-        let mut block_instructions = vec![];
-        match stmt {
-            ast::Stmt::Null => {}
-            ast::Stmt::Return(Some(expr)) => {
-                let Expr {
-                    mut instructions,
-                    val,
-                } = Expr::parse_with_and_convert(expr, symbols, make_temp_var);
-                instructions.push(Instruction::Return(Some(val)));
-                block_instructions.extend(instructions);
-            }
-            ast::Stmt::Return(None) => {
-                block_instructions.push(Instruction::Return(None));
-            }
-            ast::Stmt::Expr(expr) => {
-                let Expr { instructions, .. } =
-                    Expr::parse_with_and_convert(expr, symbols, make_temp_var);
-                block_instructions.extend(instructions);
-            }
-            ast::Stmt::Compound(block) => {
-                block_instructions.extend(Self::parse_block_with(block, symbols, make_temp_var));
-            }
-            ast::Stmt::Goto(label) => {
-                block_instructions.push(Instruction::Jump(label));
-            }
-            ast::Stmt::Label { name, stmt } => {
-                block_instructions.push(Instruction::Label(name));
-                block_instructions.extend(Self::parse_stmt_with(*stmt, symbols, make_temp_var));
-            }
-            ast::Stmt::Break(label) => {
-                let label = Rc::new(format!("{}.break", label.unwrap()));
-                block_instructions.push(Instruction::Jump(label));
-            }
-            ast::Stmt::Continue(label) => {
-                let label = Rc::new(format!("{}.continue", label.unwrap()));
-                block_instructions.push(Instruction::Jump(label));
-            }
-            ast::Stmt::While {
-                condition,
-                body,
-                label,
-            } => {
-                let label = label.unwrap();
-                let label_continue = Rc::new(format!("{label}.continue"));
-                let label_break = Rc::new(format!("{label}.break"));
-
-                // Label(continue_label)
-                block_instructions.push(Instruction::Label(Rc::clone(&label_continue)));
-                // <instructions for condition>
-                // v = <result of condition>
-                let Expr { instructions, val } =
-                    Expr::parse_with_and_convert(condition, symbols, make_temp_var);
-                block_instructions.extend(instructions);
-                // JumpIfZero(v, break_label)
-                block_instructions.push(Self::JumpIfZero {
-                    condition: val,
-                    target: Rc::clone(&label_break),
-                });
-                // <instructions for body>
-                block_instructions.extend(Instruction::parse_stmt_with(
-                    *body,
-                    symbols,
-                    make_temp_var,
-                ));
-                // Jump(continue_label)
-                block_instructions.push(Instruction::Jump(label_continue));
-                // Label(break_label)
-                block_instructions.push(Instruction::Label(label_break));
-            }
-            ast::Stmt::DoWhile {
-                body,
-                condition,
-                label,
-            } => {
-                let label = label.unwrap();
-                let label_start = Rc::new(format!("{label}.start"));
-                let label_continue = Rc::new(format!("{label}.continue"));
-                let label_break = Rc::new(format!("{label}.break"));
-
-                // Label(start)
-                block_instructions.push(Instruction::Label(Rc::clone(&label_start)));
-                // <instructions for body>
-                block_instructions.extend(Instruction::parse_stmt_with(
-                    *body,
-                    symbols,
-                    make_temp_var,
-                ));
-                // Label(continue_label)
-                block_instructions.push(Instruction::Label(label_continue));
-                // <instructions for condition>
-                // v = <result of condition>
-                let Expr { instructions, val } =
-                    Expr::parse_with_and_convert(condition, symbols, make_temp_var);
-                block_instructions.extend(instructions);
-                // JumpIfNotZero(v, start)
-                block_instructions.push(Instruction::JumpIfNotZero {
-                    condition: val,
-                    target: label_start,
-                });
-                // Label(break_label)
-                block_instructions.push(Instruction::Label(label_break));
-            }
-            ast::Stmt::For {
-                init,
-                condition,
-                post,
-                body,
-                label,
-            } => {
-                let label = label.unwrap();
-                let label_start = Rc::new(format!("{label}.start"));
-                let label_continue = Rc::new(format!("{label}.continue"));
-                let label_break = Rc::new(format!("{label}.break"));
-
-                // <instructions for init>
-                match *init {
-                    ast::ForInit::Decl(decl) => {
-                        block_instructions.extend(Instruction::parse_var_decl_with(
-                            decl,
-                            symbols,
-                            make_temp_var,
-                        ));
-                    }
-                    ast::ForInit::Expr(Some(expr)) => {
-                        let Expr { instructions, .. } =
-                            Expr::parse_with_and_convert(expr, symbols, make_temp_var);
-                        block_instructions.extend(instructions);
-                    }
-                    ast::ForInit::Expr(None) => {}
-                }
-                // Label(start)
-                block_instructions.push(Instruction::Label(Rc::clone(&label_start)));
-                // <instructions for condition>
-                // v = <result of condition>
-                if let Some(cond) = condition {
-                    let Expr {
-                        instructions: instructions_condition,
-                        val: val_condition,
-                    } = Expr::parse_with_and_convert(cond, symbols, make_temp_var);
-                    block_instructions.extend(instructions_condition);
-                    // JumpIfZero(v, break_label)
-                    block_instructions.push(Instruction::JumpIfZero {
-                        condition: val_condition,
-                        target: Rc::clone(&label_break),
-                    });
-                }
-                // <instructions for body>
-                block_instructions.extend(Instruction::parse_stmt_with(
-                    *body,
-                    symbols,
-                    make_temp_var,
-                ));
-                // Label(continue_label)
-                block_instructions.push(Instruction::Label(label_continue));
-                // <instructions for post>
-                if let Some(post) = post {
-                    let Expr {
-                        instructions: instructions_post,
-                        ..
-                    } = Expr::parse_with_and_convert(post, symbols, make_temp_var);
-                    block_instructions.extend(instructions_post);
-                }
-                // Jump(Start)
-                block_instructions.push(Instruction::Jump(label_start));
-                // Label(break_label)
-                block_instructions.push(Instruction::Label(label_break));
-            }
-            ast::Stmt::If {
-                condition,
-                then,
-                r#else,
-            } => {
-                let (else_label, end_label) = {
-                    let label = make_temp_var();
-                    // This isn't needed and can be simplified... To Bad!
-                    let Some((name, count)) = label.as_str().split_once('.') else {
-                        unreachable!("label should always be name.count");
-                    };
-                    let else_label = format!("{name}.{count}.if_else");
-                    let end_label = format!("{name}.{count}.if_end");
-                    (Rc::new(else_label), Rc::new(end_label))
-                };
-                let Expr {
-                    mut instructions,
-                    val,
-                } = Expr::parse_with_and_convert(condition, symbols, make_temp_var);
-
-                instructions.push(Self::JumpIfZero {
-                    condition: val,
-                    target: Rc::clone(match r#else {
-                        Some(_) => &else_label,
-                        None => &end_label,
-                    }),
-                });
-
-                instructions.extend(Instruction::parse_block_with(
-                    ast::Block(vec![ast::BlockItem::Stmt(*then)]),
-                    symbols,
-                    make_temp_var,
-                ));
-
-                if let Some(r#else) = r#else {
-                    instructions.push(Instruction::Jump(Rc::clone(&end_label)));
-                    instructions.push(Instruction::Label(Rc::clone(&else_label)));
-                    instructions.extend(Instruction::parse_block_with(
-                        ast::Block(vec![ast::BlockItem::Stmt(*r#else)]),
-                        symbols,
-                        make_temp_var,
-                    ));
-                }
-
-                instructions.push(Instruction::Label(Rc::clone(&end_label)));
-                block_instructions.extend(instructions);
-            }
-            ast::Stmt::Case {
-                value: _,
-                stmt,
-                label,
-            } => {
-                let label = label.expect("Case must have label");
-                block_instructions.push(Instruction::Label(Rc::clone(&label)));
-                block_instructions.extend(Self::parse_stmt_with(*stmt, symbols, make_temp_var));
-            }
-            ast::Stmt::Default { label, stmt } => {
-                let label = label.expect("Default must have label");
-                block_instructions.push(Instruction::Label(Rc::clone(&label)));
-                block_instructions.extend(Self::parse_stmt_with(*stmt, symbols, make_temp_var));
-            }
-            ast::Stmt::Switch {
-                condition,
-                body,
-                label,
-                cases,
-                default,
-            } => {
-                let label = label.expect("Switch statement must be labelled.");
-                let cases = cases.expect("Cases must have been populated.");
-                let break_label = Rc::new(format!("{label}.break"));
-                // Cases:
-                // 1. There are no cases: Still evaluate the condition and make
-                //  instructions in case a goto jumps to a label in them.
-                // 2. The switch value is a literal (or has been const-evaled):
-                //  Perform a compile time comparison to all the cases and make
-                //  a single jump to the right location.
-                // 3. Perform a linear comparison for each case and jump if the
-                //  value matches.
-                if cases.is_empty() {
-                    let Expr { instructions, .. } =
-                        Expr::parse_with_and_convert(condition, symbols, make_temp_var);
-                    block_instructions.extend(instructions);
-                    block_instructions.push(Instruction::Jump(
-                        default.unwrap_or(Rc::clone(&break_label)),
-                    ));
-                    block_instructions.extend(Self::parse_stmt_with(*body, symbols, make_temp_var));
-                } else if let ast::Expr::Constant(cond) = condition {
-                    let jump_label = cases
-                        .iter()
-                        .find(|&&(case, _)| case == cond)
-                        .map_or(default, |(_, label)| Some(Rc::clone(label)));
-                    let has_jump_label = jump_label.is_some();
-                    let jump_label = jump_label.unwrap_or(break_label.clone());
-                    block_instructions.push(Instruction::Jump(jump_label.clone()));
-                    block_instructions.extend(Self::parse_stmt_with(*body, symbols, make_temp_var));
-                    if !has_jump_label {
-                        block_instructions.push(Instruction::Label(jump_label));
-                    }
-                } else {
-                    let Expr {
-                        instructions,
-                        val: switch_val,
-                    } = Expr::parse_with_and_convert(condition, symbols, make_temp_var);
-                    block_instructions.extend(instructions);
-                    for (case, label) in cases.iter() {
-                        let Expr {
-                            instructions,
-                            val: case_val,
-                        } = Expr::parse_with_and_convert(
-                            ast::Expr::Constant(*case),
-                            symbols,
-                            make_temp_var,
-                        );
-                        block_instructions.extend(instructions);
-                        let dst = Function::make_tacky_temp_var(
-                            ast::Type::bool(),
-                            symbols,
-                            make_temp_var,
-                        );
-                        block_instructions.push(Self::Binary {
-                            op: BinaryOp::Equal,
-                            src1: switch_val.clone(),
-                            src2: case_val,
-                            dst: dst.clone(),
-                        });
-                        block_instructions.push(Self::JumpIfNotZero {
-                            condition: dst,
-                            target: Rc::clone(label),
-                        });
-                    }
-                    block_instructions.push(Instruction::Jump(
-                        default.unwrap_or(Rc::clone(&break_label)),
-                    ));
-                    block_instructions.extend(Self::parse_stmt_with(*body, symbols, make_temp_var));
-                }
-                // Break label always goes after all instructions
-                block_instructions.push(Instruction::Label(break_label));
-            }
-        }
-        block_instructions
-    }
-    fn parse_block_with(
-        node: ast::Block,
-        symbols: &mut SymbolTable,
-        make_temp_var: &mut impl FnMut() -> String,
-    ) -> Vec<Self> {
-        let mut block_instructions = vec![];
-        for item in node.into_items().into_iter() {
-            match item {
-                // Statics already get initialized at the top level.
-                // If we reinitialized them here they would act like local
-                // variables (suboptimal)
-                ast::BlockItem::Decl(ast::Declaration::VarDecl(ast::VarDecl {
-                    storage_class: Some(ast::StorageClass::Static),
-                    ..
-                })) => {}
-                ast::BlockItem::Decl(decl) => {
-                    block_instructions.extend(Self::parse_decl_with(decl, symbols, make_temp_var));
-                }
-                ast::BlockItem::Stmt(stmt) => {
-                    block_instructions.extend(Self::parse_stmt_with(stmt, symbols, make_temp_var));
-                }
-            }
-        }
-        block_instructions
-    }
-}
+use super::*;
 
 #[derive(Debug, PartialEq)]
 pub struct Expr {
-    instructions: Vec<Instruction>,
-    val: Val,
+    pub instructions: Vec<Instruction>,
+    pub val: Val,
 }
 
 impl Expr {
@@ -1691,7 +926,7 @@ impl Expr {
         }
     }
 
-    fn parse_with_and_convert(
+    pub(crate) fn parse_with_and_convert(
         node: ast::Expr,
         symbols: &mut SymbolTable,
         make_temp_var: &mut impl FnMut() -> String,
@@ -1718,206 +953,5 @@ impl Expr {
                 }
             }
         }
-    }
-}
-
-#[derive(Debug)]
-enum ExprResult {
-    PlainOperand(Expr),
-    DerefrencedPointer(Expr),
-}
-
-#[derive(Clone, Debug, PartialEq)]
-pub enum Val {
-    Constant(ast::Constant),
-    Var(Rc<String>),
-}
-
-impl Val {
-    pub fn get_type(&self, symbols: &SymbolTable) -> ast::Type {
-        match self {
-            Self::Constant(c) => c.get_type(),
-            Self::Var(name) => {
-                let Some(entry) = symbols.get(name) else {
-                    unreachable!("Variable name '{name}' not found in symbol table");
-                };
-                entry.r#type.clone()
-            }
-        }
-    }
-}
-
-impl From<ast::Constant> for Val {
-    fn from(node: ast::Constant) -> Self {
-        Self::Constant(node)
-    }
-}
-
-#[derive(Clone, Debug, PartialEq)]
-pub enum UnaryOp {
-    Negate,
-    Complement,
-    Not,
-}
-
-impl From<ast::UnaryOp> for UnaryOp {
-    fn from(node: ast::UnaryOp) -> Self {
-        match node {
-            ast::UnaryOp::Complement => Self::Complement,
-            ast::UnaryOp::Negate => Self::Negate,
-            ast::UnaryOp::Not => Self::Not,
-            _ => unreachable!(),
-        }
-    }
-}
-
-#[derive(Clone, Copy, Debug, PartialEq)]
-pub enum BinaryOp {
-    Add,
-    Subtract,
-    Multiply,
-    Divide,
-    Remainder,
-    BitAnd,
-    BitOr,
-    Xor,
-    LShift,
-    RShift,
-    Equal,
-    NotEqual,
-    LessThan,
-    LessOrEqual,
-    GreaterThan,
-    GreaterOrEqual,
-    Assign,
-    AddAssign,
-    SubAssign,
-    MultAssign,
-    DivAssign,
-    ModAssign,
-    AndAssign,
-    OrAssign,
-    XorAssign,
-    LShiftAssign,
-    RShiftAssign,
-}
-
-impl From<ast::BinaryOp> for BinaryOp {
-    fn from(node: ast::BinaryOp) -> Self {
-        match node {
-            ast::BinaryOp::Add => Self::Add,
-            ast::BinaryOp::Subtract => Self::Subtract,
-            ast::BinaryOp::Multiply => Self::Multiply,
-            ast::BinaryOp::Divide => Self::Divide,
-            ast::BinaryOp::Remainder => Self::Remainder,
-            ast::BinaryOp::BitAnd => Self::BitAnd,
-            ast::BinaryOp::BitOr => Self::BitOr,
-            ast::BinaryOp::Xor => Self::Xor,
-            ast::BinaryOp::LShift => Self::LShift,
-            ast::BinaryOp::RShift => Self::RShift,
-            ast::BinaryOp::And => unreachable!("Logical operands are not Binary operands"),
-            ast::BinaryOp::Or => unreachable!("Logical operands are not Binary operands"),
-            ast::BinaryOp::Equal => Self::Equal,
-            ast::BinaryOp::NotEqual => Self::NotEqual,
-            ast::BinaryOp::LessThan => Self::LessThan,
-            ast::BinaryOp::LessOrEqual => Self::LessOrEqual,
-            ast::BinaryOp::GreaterThan => Self::GreaterThan,
-            ast::BinaryOp::GreaterOrEqual => Self::GreaterOrEqual,
-            ast::BinaryOp::Assign => Self::Assign,
-            ast::BinaryOp::AddAssign => Self::AddAssign,
-            ast::BinaryOp::SubAssign => Self::SubAssign,
-            ast::BinaryOp::MultAssign => Self::MultAssign,
-            ast::BinaryOp::DivAssign => Self::DivAssign,
-            ast::BinaryOp::ModAssign => Self::ModAssign,
-            ast::BinaryOp::AndAssign => Self::AndAssign,
-            ast::BinaryOp::OrAssign => Self::OrAssign,
-            ast::BinaryOp::XorAssign => Self::XorAssign,
-            ast::BinaryOp::LShiftAssign => Self::LShiftAssign,
-            ast::BinaryOp::RShiftAssign => Self::RShiftAssign,
-            ast::BinaryOp::Ternary => {
-                unreachable!("Ternary expressions are not true Binary operands")
-            }
-        }
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_return_literal() {
-        let mut symbols = SymbolTable::default();
-        let ast = ast::Block(vec![ast::BlockItem::Stmt(ast::Stmt::Return(Some(
-            ast::Expr::Constant(ast::Constant::I32(2)),
-        )))]);
-        let mut counter = 0;
-        let mut make_temp_var = Function::make_temp_var(Rc::new("test".to_string()), &mut counter);
-        let actual = Instruction::parse_block_with(ast, &mut symbols, &mut make_temp_var);
-        let expected = vec![Instruction::Return(Some(Val::Constant(
-            ast::Constant::I32(2),
-        )))];
-        assert_eq!(actual, expected);
-    }
-
-    #[test]
-    fn test_return_unary() {
-        let mut symbols = SymbolTable::default();
-        let ast = ast::Block(vec![ast::BlockItem::Stmt(ast::Stmt::Return(Some(
-            ast::Expr::Unary {
-                op: ast::UnaryOp::Complement,
-                expr: Box::new(ast::Expr::Constant(ast::Constant::I32(2))),
-            },
-        )))]);
-        let mut counter = 0;
-        let mut make_temp_var = Function::make_temp_var(Rc::new("test".to_string()), &mut counter);
-        let actual = Instruction::parse_block_with(ast, &mut symbols, &mut make_temp_var);
-        let expected = vec![
-            Instruction::Unary {
-                op: UnaryOp::Complement,
-                src: Val::Constant(ast::Constant::I32(2)),
-                dst: Val::Var("tacky.test.0".to_string().into()),
-            },
-            Instruction::Return(Some(Val::Var("tacky.test.0".to_string().into()))),
-        ];
-        assert_eq!(actual, expected);
-    }
-    #[test]
-    fn test_return_nested_unary() {
-        let mut symbols = SymbolTable::default();
-        let ast = ast::Block(vec![ast::BlockItem::Stmt(ast::Stmt::Return(Some(
-            ast::Expr::Unary {
-                op: ast::UnaryOp::Negate,
-                expr: Box::new(ast::Expr::Unary {
-                    op: ast::UnaryOp::Complement,
-                    expr: Box::new(ast::Expr::Unary {
-                        op: ast::UnaryOp::Negate,
-                        expr: Box::new(ast::Expr::Constant(ast::Constant::I32(2))),
-                    }),
-                }),
-            },
-        )))]);
-        let mut counter = 0;
-        let mut make_temp_var = Function::make_temp_var(Rc::new("test".to_string()), &mut counter);
-        let actual = Instruction::parse_block_with(ast, &mut symbols, &mut make_temp_var);
-        let expected = vec![
-            Instruction::Unary {
-                op: UnaryOp::Negate,
-                src: Val::Constant(ast::Constant::I32(2)),
-                dst: Val::Var("tacky.test.0".to_string().into()),
-            },
-            Instruction::Unary {
-                op: UnaryOp::Complement,
-                src: Val::Var("tacky.test.0".to_string().into()),
-                dst: Val::Var("tacky.test.1".to_string().into()),
-            },
-            Instruction::Unary {
-                op: UnaryOp::Negate,
-                src: Val::Var("tacky.test.1".to_string().into()),
-                dst: Val::Var("tacky.test.2".to_string().into()),
-            },
-            Instruction::Return(Some(Val::Var("tacky.test.2".to_string().into()))),
-        ];
-        assert_eq!(actual, expected);
     }
 }
